fixup! refactor(TaskRunDetailsView): update to use useTaskRunV2 hook
